Add "favoriteicecream" attribute
The purpose of adding this attribute is to see whether or not the plugin
can be modified once the IDM server(s) is/are live and have the plugin
installed. It would be good to be able to modify the plugin in the
future if there are ever any fields that we'd like to add to or remove
from it, but if that's not possible, we might want to sit down and
consider what else the plugin should have (if anything) before going to
production.